lab1/flaskapp/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.2
matplotlib==3.9.2
numpy==2.1.2
Pillow==11.0.0
Requests==2.32.3
gunicorn==23.0.0